Various optimizations in an attempt to reduce the overhead of the SEATSWorker. This commit more than doubles the speed of SEATS but that still only puts us at 50 txn/s. Will continue with this tomorow.
git-svn-id: https://oltpbenchmark.googlecode.com/svn/trunk@328 8c9255d6-0b70-036a-fcb6-8a999947b34c

+++ src/com/oltpbenchmark/util/Histogram.java
@@ -177,7 +177,7 @@ public class Histogram<X> implements JSONSerializable {
         // If we already have this value in our histogram, then add the new count
         // to its existing total
         if (this.histogram.containsKey(value)) {
-            count += this.histogram.get(value);
+            count += this.histogram.get(value).longValue();
         } else if (this.histogram.isEmpty()) {
             this.min_count = count;
             this.min_count_values.add(value);
